requirements_current.txt
accelerate==1.10.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
altair==5.5.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
arize-phoenix==12.9.0
arize-phoenix-client==1.21.0
arize-phoenix-evals==2.5.0
arize-phoenix-otel==0.13.1
async-timeout==4.0.3
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.3.0
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
effdet==0.4.1
email-validator==2.3.0
emoji==2.15.0
exceptiongroup==1.3.0
fastapi==0.121.0
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
galileo==1.28.0
galileo-core==3.67.3
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-auth==2.42.1
google-cloud-vision==3.11.0
googleapis-common-protos==1.71.0
graphql-core==3.2.7
greenlet==3.2.4
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
kubernetes==34.1.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-pinecone==0.2.13
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.10.1
networkx==3.2.1
nltk==3.9.2
numpy==2.0.1
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.19.2
openai==2.6.1
opencv-python==4.12.0.88
openinference-instrumentation==0.1.41
openinference-instrumentation-langchain==0.1.54
openinference-instrumentation-openai==0.1.39
openinference-semantic-conventions==0.1.24
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-openai==0.47.5
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
ply==3.11
posthog==5.4.0
prometheus_client==0.23.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlean.py==3.50.4.5
starlette==0.49.3
strawberry-graphql==0.270.1
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.21
tokenizers==0.22.1
toml==0.10.2
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.3
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.21.0
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
